+++ test/repository.spec.js
@@ -6,6 +6,7 @@ import {WithCustomRepository} from './resources/entity/with-custom-repository';
 import {SimpleCustom} from './resources/repository/simple-custom';
 import {Container} from 'aurelia-dependency-injection';
 import {Foo} from './resources/entity/foo';
+import {WithType} from './resources/entity/with-type';
 import {Custom} from './resources/entity/custom';
 import {Config} from 'spoonx/aurelia-api';
 
@@ -36,7 +37,7 @@ function getEntityManager(container) {
   container         = container || getContainer();
   let entityManager = new EntityManager(container);
 
-  return entityManager.registerEntities([WithResource, Foo, Custom, WithCustomRepository, WithAssociations]);
+  return entityManager.registerEntities([WithResource, Foo, Custom, WithCustomRepository, WithAssociations, WithType]);
 }
 
 describe('Repository', function() {
@@ -192,6 +193,25 @@ describe('Repository', function() {
       expect(populated instanceof Entity).toBe(true);
       expect(populated.isClean()).toBe(true);
       expect(populated.isDirty()).toBe(false);
+    });
+
+    it('Should properly cast values when defined in the entity', function() {
+      var repository = constructRepository('with-type'),
+          populated  = repository.getPopulatedEntity({
+            created : '2016-02-22T18:00:00.000Z',
+            disabled: '0',
+            age     : '24',
+            titanic : '500'
+          });
+
+      expect(populated instanceof Entity).toBe(true);
+      expect(populated.created).toEqual(new Date('2016-02-22T18:00:00.000Z'));
+      expect(populated.asObject()).toEqual({
+        created : new Date('2016-02-22T18:00:00.000Z'),
+        disabled: false,
+        age     : 24,
+        titanic : 500.00
+      });
     });
 
     it('Should return a populated instance with associations.', function() {