+++ test/resources/entity/with-type.js
@@ -1,0 +1,17 @@
+import {resource, type} from '../../../src/index';
+import {Entity} from '../../../src/entity';
+
+@resource('with-type')
+export class WithType extends Entity {
+  @type('datetime')
+  created = null;
+
+  @type('boolean')
+  disabled = null;
+
+  @type('integer')
+  age = null;
+
+  @type('float')
+  titanic = null;
+}