+++ doc/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.1.0 (2016-02-23)
+
+
+#### Features
+
+* **decorator:** Added type decorator ([3e70062e](https://github.com/SpoonX/aurelia-orm/commit/3e70062e0f11d82bdadc80985a1eba625a152182))
+* **project:** Expose type decorator from project index ([e20459cc](https://github.com/SpoonX/aurelia-orm/commit/e20459cca82ffa2b5196ec1bcf03f0487b29a591))
+* **repository:** Cast values if they have type decorators ([a40d7632](https://github.com/SpoonX/aurelia-orm/commit/a40d7632f856a78ff78c0428075c78f90af67a68))
+
+
 <a name"2.0.3"></a>
 ## 2.0.3 (2016-01-25)